Rework rendering engine into a more sensible form using RenderBatcher

+++ src/Strawberry/UI/Rendering/Text/Glyph.hpp
@@ -44,7 +44,7 @@ namespace Strawberry::UI
 		/// This glyph's contours.
 		ContourList mContours;
 
-		Core::Math::Vec2i mAdvance;
-		Core::Math::Vec2i mHorizontalBearing;
+		Core::Math::Vec2f mAdvance;
+		Core::Math::Vec2f mHorizontalBearing;
 	};
 }
